+++ src/LocalFile.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace danog\MadelineProto;
+
+/**
+ * Indicates a local file to upload.
+ */
+final class LocalFile
+{
+    public function __construct(
+        public readonly string $file
+    ) {
+    }
+}